01_Prompt Engineering and Prototyping Best Practices/pyproject.toml
[project]
name = "aie-s1-a1"
version = "0.1.0"
description = "AI Engineering Bootcamp - Session 1 - Activity 1 - Advanced"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit==0.7.700",
    "cohere==4.37",
    "openai==1.3.5",
    "pydantic==2.10.1",
    "python-dotenv==1.0.0",
    "tiktoken== 0.9.0",
]
